requirements.txt
Django>=5.0
oracledb>=3.0.0
python-decouple>=3.8
Pillow>=10.0.0
